requirements.txt
streamlit
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-core
python-dotenv
faiss-cpu
sentence-transformers